# Massively speedup ContainerRegistry.findContainers()