Only use regular expressions when we need to. Most of the time simple string comparison is all that is needed.

This cuts the loading time of Cura by ~35% for me (from 43 to 34 seconds on my machine). Specifically, and in combination with the optimisation in https://github.com/Ultimaker/Cura/commit/30a679ddf9019aad721e8b1deccd7e453a216683 this has cut deserialisation of the XMLMaterialProfiles from 24 seconds to 4 seconds.

I am putting this in a PR because I am not 100% sure it does not break things; it needs testing, and I am curious to know if it speeds things up on other systems as dramatically as it does on mine.

CURA-2187